Improve code organization
To make sure it is still readable after upcoming changes to helmfile
