+++ pkg/app/two_pass_renderer_test.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"fmt"
